Move block factories into factories/ folder
