+++ packages/api/cms-api/src/blocks/factories/createTextImageBlock.ts
@@ -1,8 +1,8 @@
 import { Type } from "class-transformer";
 import { IsEnum, IsString } from "class-validator";
 
-import { PixelImageBlock } from "../dam/blocks/pixel-image.block";
-import { IsAllowedImageAspectRatio } from "../dam/images/validators/is-allowed-aspect-ratio.validator";
+import { PixelImageBlock } from "../../dam/blocks/pixel-image.block";
+import { IsAllowedImageAspectRatio } from "../../dam/images/validators/is-allowed-aspect-ratio.validator";
 import {
     Block,
     BlockData,
@@ -13,10 +13,10 @@ import {
     ExtractBlockInput,
     inputToData,
     SimpleBlockInputInterface,
-} from "./block";
-import { ChildBlock } from "./decorators/child-block";
-import { ChildBlockInput } from "./decorators/child-block-input";
-import { BlockField } from "./decorators/field";
+} from "../block";
+import { ChildBlock } from "../decorators/child-block";
+import { ChildBlockInput } from "../decorators/child-block-input";
+import { BlockField } from "../decorators/field";
 
 export enum ImagePosition {
     Left = "left",
